Require firmware signatures and simplify verification
This addresses #88 by simplifying validation and requiring signatures on
all .fw files. Making this changed removed code that broke when what was
intended to be a human-read text message from fwup was modified.
Requiring firmware signatures is also a best practice for firmware
update distribution and it prevents any shenanigans on .fw files while
their in the possession of Nerves Hub.

`fwup` doesn't provide enough machine-readable granulatary into why an
archive doesn't pass validation to give the user some help as to what
went wrong if validation fails. It looks possible to enhance `fwup` to
help out in this area should that be needed.