Support email addresses with capital letters
The authentication code previously downcased email addresses. Since the
email address downcase was only done in one place, this location was
inconsistent and it caused authentication failures.

This addresses https://github.com/nerves-hub/nerves_hub_cli/issues/113
and #423. The latter issue talks about normalizing email addresses
throughout but mentions edge cases in the RFCs that appear to make this
non-trivial. This change punts on normalization. If the user wants a
capital letter or something else odd, but legal, in their email address,
they can have it.