fix delete_file/1 in Upload.File
This just fixes a simple edge case of attempting to delete a firmware file stored locally
that doesn't exist anymore. Really, this only affects running NervesHub locally (or anyone
else who might use this file store instead of S3). In the case of local development, fw
files are stored in tmp locations so a restart could blow it away while the record exists
in the app. Then attempting to delete the record would fail. This just checks if it is there
first before attempting to hard delete