Tell org users about the addition or removal of a new user
Issue #403
When a user accepts an invitation or when a new user is added to or
removed from an organization (either via the ClI or UI) generate and send emails to
every user in the organization.

This event is a result of an action in the nerves_hub_api and
nerves_hub_www apps.