Dont attempt to alter table data within UniqueDeploymentNameProductId migration
With recent changes to `Deployment` schema, it breaks when trying to run this migration in the CI.
Which means it will also break for new users attempting to create the DB from start.

The error is caused by an old migration expecting an old definition of `Deployment` so it can use it
while running to fill in data. So this removes that with the expectation that we can create a separate
task to run to fix existing DB data if still needed