Fix devices#index status weirdness
Turns out when I moved the devices#index page to LiveView, I forgot to remote the js
code connecting to a socket and watching for presence change events. So every now and
then there was a race condition between the JS updating devices statuses on the page
and LiveView.

So to fix, :boom: the socket javascript!