All ANSI \e[22m in console output
Some messaages from remote iex session come in with an escape of `\e[22m` and it would
break the rendering. This parameter is [`Normal color or intensity`](https://en.wikipedia.org/wiki/ANSI_escape_code#SGR_parameters) so I think just ignoring it and defaulting to the HTML coloring is fine.