Scope failure queries to firmware UUID
Fixes #474

Adds `firmware_uuid` to an Audit Log event created when a device is told to update.
Then when we use the AuditLog to check failure rates and thresholds, we check for that UUID
as well to ensure we're correctly matching failures. This allows reusuing deployments with
different firmwares and adds little more accuracy to device failure checking.