Support remote iex lines that are lists
A recent change in https://github.com/nerves-hub/nerves_hub/pull/102 added support to handle
io_requests requiring a function to be run to format the data. In some cases, that ends up
being a list of characters or numbers (representing a character) that would break the console
session because to regex is used to generate the ANSI HTML and it couldn't handle the list.

So this just calls `to_string/1` on data that comes in from the device before attempting to convert
to HTML and display.