Fix audit! failures on broadcast
We recently changed update message audits to inject the firmware uuid in it. However,
on a broadcast we were not sending the full deployment struct and so we could not lookup
the firmware (or ensure it was there) to perform the audit. So, this just changes a few
things to send that full struct to the channel, but remove it before sending to the device.

The channel will also do a lazy lookup to search based on deployment id if for some reason
the deployment struct is not in the broadcast payload. Also add a couple regression tests so
this should be backwards compatible (though I didn't find anyone actually using other keys
in the payload besides `firmware_url`)