Fix deleting device from UI
Attempting to delete a device from the UI would error because LiveView handles CSRF token a little different and it would not match the token expected for `device#delete` endpoint in the device controller.

So this adjusts that to handle the LiveView session for DeviceLive.Index a little differently by going back to the original idea of forwarding from the controller and passing the needed csrf token.

This can probably be adjusted if needed once https://github.com/phoenixframework/phoenix_live_view/issues/111 is resolved.