Add catch-all for LiveView mounts
When we change anything about the session structure of a LiveView, it causes tons of rollbar errors after deployment because any existing browser tabs/windows are trying to reuse the old session structure which no longer has a matching `mount/2` clause for it.

So, this adds a catch-all to redirect to the home path if that apps.