Support full terminal emulation in remote IEx console
This adds support for using xterm.js as the web terminal emulation for
the remote IEx console. It simplifies things a bit and allows the fun
terminal features we all know and love like tab completion, history search,
arrow movement, etc.

To use this new web console, a device will need to update `nerves_hub_link`.

For now, both the LiveView terminal and this will be supported and routing
happens automatically based on what the device supports