Find CA by ski and not aki
The current code assumes that all ca certificates are root certificates by searching for on who has the same authority and subject identifier (a self signed root cert) This behavior is wrong and prevents intermeduates from being used for signing. Updating the lookup to find the ca by ski will allow for both self signed and intermediates to be used.