Change domain shown in emails
When a user is created or requests a password change, the email says its
an action on `nerves-hub.org`. However, if this is self hosted, that may
not actually be the domain.

This changes to use the `EmailView.base_url()` instead when referencing
where the email was generated which pulls from Application config
variables that may have been customized in a self-hosted instance.