Increase download URL valid time to 24 hours
Firmware download URLs sent to devices are currently valid for 10
minutes. Devices that can restart downloads need the URL to be valid for
the duration of the download. This causes an issue for devices
downloading over slow cellular links that may not be able to complete
the firmware update for an hour or two.

This changes the URL valid time to be 24 hours. This should be more than
long enough for any device to download a firmware update and if it takes
a device longer than this, requiring some special case handling to fetch
a new URL seems reasonable.

Lengthening the URL valid time only changes the signature, so this doesn't
have side effects like preventing an administrator from deleting a
firmware on NervesHub. Security-wise, the signed URL is sent over the
encrypted channel or HTTPS connection, so if that's compromised, then
the firmware could be downloaded by an unauthorized party. Presumably
anyone with this ability would be able to download the firmware within
10 minutes as well.