Remove invalid update requests from audit log
The device quarantine logic uses the audit log to determine whether
a device is misbehaving. One sign of a misbehaving device is too
many update requests for the same firmware version.

This fixes an audit record on 'join' events that was unconditionally
written even if the update request was currently being processed
by the device. As a result, devices with flaky network connections
won't get quarantined so long as they report that they're downloading
the right firmware on reconnect.

This adds a check for the audit log on `join` which will check
if a device is currently already downloading an update, which
will prevent it from being quarantined when a channel connecion
is dropped mid download.