add Just In Time Provisioning for Devices
This follows as closely as possible to AWS IOT. Creates a new
table called `jitp` which stores the default tags and description
for devices provisioned with JITP.

Upon connect, NervesHub checks if the product is configured to
have JITP enabled. if enabled, it will copy the defaults and create
a new Device record.