Ignore SSL auth when device has been soft deleted
Right now if a device is deleted, it is a soft delete. When it
attempts to connect again, SSL verification will succeed but the
device will not show anywhere in the UI.

This instead fixes that to prevent authentication for devices that are
deleted and require the user to explicitly restore it.