Account for expired CA certificates in requests
With OTP 24, an expired CA in the chain needs to be handled explicitly. In our case,
an expired CA may still be presented in the chain for already registered device certs
and CA certs in which we should still allow the connection.

This change adds a lookup for the expired certificate to see if it is a CA we
already know about then allows the connection if it is. This should be safe as
any attempts to use this expired CA will still fail later one once the new device
certificate is presented and goes through validation