Implement user access token authentication
This is heavily inspired by [GitHub authentication token formats](https://github.blog/2021-04-05-behind-githubs-new-authentication-token-formats/).

There are a few goals in mind with this PR:
* Remove the requirement for `nerves_hub_ca` (User API authentication is the only
  thing that is really relying on it)
* Make CLI interaction easier to manage for the user
* Make CLI easier to use in a CI or automated context

The token that is currently created will undergo a CRC check before attempting
to hit the DB in order to return many auth requests from modified strings.

You can created tokens in the web UI for now. Though this also includes
the pieces to allow the API to implement the change as well