Set TLS version to 1.2 for the server
OTP >= 24.3 made changes to the TLS 1.3 implementation and now prefers it
as the default. Some of the hardended rules break the custom client authentication
setup of NervesHub in slightly confusing ways.

This sets the expected server version to TLS1.2 to hopefully reduce these
errors and make things a bit more consistent and predictable for now
until able to implement #852 and other changes with later OTP releases.