Force TLS 1.2 for legacy firmware devices
Devices using OTP < 25.1 failing when attempting to negotiate a server or client that
supports both TLS 1.3 and TLS 1.2. This is a pretty damaging bug for devices that
may be deployed on old versions, or have sat dormant with an old version and the
next attempt to connect will be broken.

So for now we will enforce TLS 1.2.

See https://github.com/erlang/otp/issues/6492#issuecomment-1323874205

Fixes NH-37