Downgrade to Elixir 1.13.2 / OTP 23.0.4
Continuation of #71
Fixes NH-43

Even _more_ investigation showed that devices using Trust'n'Go chips
which supplied a signer cert would be rejected. For whatever reason, the
signer cert would not be seen as a complete chain, presented to the `NervesHubDevice.SSL.verify_fun/3`
function before the device cert, and would fail since it is unknown and
not pinned.

This behavior was introduced with OTP 23.1, so for now we will downgrade
to OTP 23.0.4 to allow more time to investigate solutions. This has also been
tested with the previous concerns of #71 to support various versions of
OTP running on devices that may present a mixture of TLS 1.2 & 1.3