Track device connected status via gproc
Instead of Phoenix Presence, which only lets us list out the entire set
of connected devices. This will eventually be a problem when we have
hundreds of thousands of devices connected and each page load is going
to copy that list into memory.

Publish connection_change updates anytime a device changes presence,
instead of relying on Presence diffs.