Fix device action removing metadata from view
When a device action (such as quarantine/unquarantine) is done on a device, only the
updated database record was being reassigned and metadata that had been added to the
view is lost. This reverts the device to show as "offline" as the device.

This fix adds the metadata from the currently assigned device back to the updated record
and reassigns it to the socket and view