Home page renders a list of organizations (#897)
Set up a plug that loads a user and their orgs, then ensure the user
exists in another plug. This way the homepage can use this without
needing to load separately.

This helps with removing the assumption that a user has a default org
with their username as the name. Session login now goes to the home page
since there is a list to pick from. The dropdown also doesn't fire off
more queries now and reuses what was already loaded.