Update xtermjs theme to make red red and blue slightly darker
The current xtermjs theme made red show up as white. This was done since
the default xtermjs red color is hard to see on a black background.
However, this also made it hard to see coloring on error log messages.

The new version of red is a touch darker than light_red. This is still
pretty bright for red so it should hopefully be easy to read.

This also adjusts blue to be a touch darker so that it can be
differentiated from light_blue. Again, it's still pretty bright for blue
to retain its readability.

Fixes #959