Remove all keys from presence tracking except for status
Preparing for a new way of tracking presence that doesn't use gproc, and
the less we store, the better this will work out. We only really need to
know that something is online for presence tracking. The rest is stored
or we can get live while it happens as a broadcast.