Migrate certain device routes to be higher up in paths
Since we have unique identifiers across the entire NH, we can skip
having the org and the product in the route. Role validation will happen
in the route actions for now, and as the rest of the controller is
filled out we can figure out a better way to load these and validate.