Support registering only an SKI of signer CA
There are some cases where an incomplete chain is (i.e. the device was signed
by an intermediary signer and the root is not in the chain). In those cases, we
fail at the incomplete intermediary before we are even able to lookup the device
certificate to allow connection. In these situations, you would need to manually
register the intermediary so we can look it up as trusted and ignore it in the
chain.

However, there are some cases where the intermediary certificate is unknown and
only obtainable from the device (like with Microchip Trust&GO crypto chips). In
those cases where the chip is preprogrammed, you would be able to get the Subject
Key Identifier (SKI) of the intermediary signer CA from the device Authority Key
Identifier (AKI) in the device cert, usually from a manifest file.

This change fixes this situation by adding the known CA cert lookup and allows
you to register just a trusted SKI when it's not feasible to obtain all the
intermediary signer CA and pre-register them