Allow known intermediary signer CA
There are some cases where the signer CA is sent on the connection request,
but it is an intermediary CA (i.e. a middle cert that was signed by anoter
Root CA). In those cases where only the intermediary is sent in the request,
the chain is incomplete and the intermediary CA is evaluated before the
device certificate.

This fails because our simple check only considers a Root CA coming in (AKI == SKI)
and then prevents the device cert from even being considered.

This change fixes this situation by adding the known CA cert lookup. In these
situations, you would need to manually register the intermediary so we can look
it up as trusted and ignore it in the chain.