Don't `CTRL-L` on console start
In OTP 26, this now clears the screen. If you (and others) are using a console
page, the next person to reload the page will wipe the screen of all connected
users with that page open.

Not sure why that was in there, so this removes that initial request so
screens don't clear!