Remove extra certificate and device lookup queries in channel
`NervesHub.SSL` preps the device existence before letting it connect socket.
By that point, we should have strong confidence we can find the device in the DB.

Previously, the DeviceSocket was _only_ querying the certificate and then doing
another query when the device joined the channel. This instead does a single
lookup on device connect and remvoes extra queries run when the channel starts