Add a buffer to the device console
When a new web console joins, it will tell the device console that it
connected and get caught up in whatever is in the circular buffer. This
has the potential to be a little strange if you join while someone is
typing, but there's a chat to ask to pause and reload the page to get
the correct state for now.