Move all runtime config into vapor
Previous to this commit configuration was split between vapor and
`System.get_env/1` calls in config/*.exs files. Configuration was also
duplicated and overridden in certain situations which made it difficult
to track what took precedence.

Where possible I've moved all configuration into vapor so it is obvious
at boot time what is and is not required.