Move device handling to `DeviceLink`
`DeviceLink` was created to be the main logic and handler for a device
in NervesHub. Previously this was done in the `DeviceChannel` and
`DeviceLink` was only handling message subscriptions.

This moves all the logic out of `DeviceChannel` into `DeviceLink`. It
is almost a complete copy over though there are some minor refactors
to deal with the switch from a socket to a GenServer.

Many new tests were also added to `device_link_test.exs` to handle
all the cases it currently supports.

This paves the way to be able to use alternate transports with NervesHub
if websockets are not preferred