Fix release migrations
Previously, release migrations were attempting to be handled in `Enum.each` without
any return checks. This means failures could be returned but were unhandled and
false positive of success could be relayed.

Instead, this updates to match the Ecto and Phoenix recommendations for dealing
with migrations in a release. See also:

* https://hexdocs.pm/ecto_sql/3.11.1/Ecto.Migrator.html#module-example-running-migrations-in-a-release
* https://hexdocs.pm/phoenix/releases.html#ecto-migrations-and-custom-commands