Remove the reconnect timer in DeviceLink
The connection between a DeviceLink and DeviceChannel is unreliable when
there is a horde of devices trying to connect. A too large percentage of
devices can think they have a DeviceLink but it terminated and didn't
terminate the DeviceChannel. If we terminate faster, this may reduce the
number of devices in this state.