Remove Device Tracker
This removes the device tracker for a much more simplified interrogation
version. Anyone that wants to know if the device is online calls
`online?` which immediately returns false but also publishes out to the
device asking if it's online. If the device is online, it publishes a
connection state change. The two pages that use this then update the
status badge to online.