separate connection change and connection check
this makes it clearing if the device is coming online, or is already online but we just needed to check.

this also allows us to decide when to reload the device, as a `connection:status: is just a status confirmation, but a `connection:change` could include more information, like a firmware change, so we need to reload the device.