Changing deployment conditions uses a queue to update devices
Create inflight deployment checks to verify a device is still valid or
is now valid for a deployment. These are processed in a single genserver
per deployment per node. This will help prevent database pool contention
since there should only be a single query per device check per node.

Devices are still checked as fast as possible and will continue to look
for more inflight checks without being told of new updates.

Because deployment updates are processed inside a transaction, any
pubsub messages must be delayed until after the transaction is
successful.

Finding the next device to check is performed with a postgres lock
inside of a transaction. This way we only process a device once.

# Conflicts:
#	lib/nerves_hub_web/channels/device_channel.ex
#	lib/nerves_hub_web/live/devices/show.ex