Add delta support to orchestrator
The primary point of this change is that it re-enables firmware delta
support.

It also tries to alleviate the need for generating a ton of unused
delta files.

It tackles this by doing something that has already been discussed
and which I think is generally agreed upon which is give the
orchestrator the job of telling a newly connected device to update.
That removes the instant update on connect. We could reintroduce a
mechanism for the orchestrator to react to new devices if we like.

When a device joins the channel it will report it's firmware version
to the orchestrator. This gives the orchestrator the chance to fire
off Oban jobs for new firmware deltas. At the same time it will store
a list of firmware versions that have been processed or are
processing. This alleviates the need for repeated lookups.

Firmware versions that are in processing are excluded from registry
lookups to distribute updates.

When deltas are done generating the orchestrator will receive an
event causing a refresh of the delta status state.