Add alternative way to get shared secret
The Copy to clipboard way hasn't worked on Chromium for Linux
in some time and this fix adds a way to show the secret as an
option.

This has been reported by users running into it and wasting
time using the wrong credentials. I've personally hit it about
a million times.