Add option for showing updated devices on the map
This adds a way to indicate if the firmware is fully up to date on
devices for the map.

It also cleans up a few performance aspects of the map:
- Don't block initial load on fetching the devices/markers
  as it can be slow for many devices
- Change query to be more specific and not fetch as many unused devices
- Change data-plumbing so that markers are sent over push_event instead
  of via attributes with extra JSON encode/decode.

The map should load significantly faster and will indicate "Loading..."
while loading if delayed.

We should further check index usage for these new queries as I think
they could be made even faster if they don't have indexes today.