Use PR number when looking for publish commit message
This fixes an issue where jobs fail when built from forks which would have
an incomplete ref. This changes to use the PR number instead which should
get the details we need.

It also allows that step to fail to continue to job since it is such a specific,
special case