Clean stale `DeviceConnection`s (#1747)
I noticed in the DB that we have several old/stale device connections. 

Upon looking into this more, I realised that when we added the
`DeviceConnection` schema and relationships, we didn't add a way to
clean the stale connections, which might linger due to unforeseen
circumstances, like the app crashing.