The `Credo` experiment (#1740)
This PR adds credo and addresses the most important issues (and some low
hanging fruit) raised when using `--strict`.

This also adds it to our CI, but it doesn't fail it for now.

I've also modified some of the settings:

- disabled `{Credo.Check.Readability.ParenthesesOnZeroArityDefs, []}` as
I don't know if I agree with this, at the very least we should discuss
and decide if we want this enforced. (this check, if enabled, would
require us to remove `()` from functions which don't accept any
arguments)
- disabled `{Credo.Check.Refactor.CondStatements, []}` as we have two
2-condition `cond` statements in `NervesHub.Extensions` which are useful
to keep around. If we want to enforce this then we might need to change
these until it makes sense switching back to using `cond` for them.
- customized `{Credo.Check.Refactor.Nesting, [max_nesting: 3]}`, the
default is 2 but we have enough 3 level nesting that I didn't feel like
going down that path right now.
- customized `Credo.Check.Warning.MissedMetadataKeyInLoggerConfig` to
tell it that we do use `:all` in `runtime.exs`
- customized `Credo.Check.Readability.ParenthesesOnZeroArityDefs` to
require parentheses (the opposite of its default)
- enabled `Credo.Check.Warning.LeakyEnvironment` to warn us if we are
passing the full environment from the parent process when running system
commands.
- enabled `Credo.Check.Readability.ImplTrue` as @jjcarstens hates it
when `@impl true` is used
- and finally, enabled `Credo.Check.Readability.MultiAlias` as the use
of `alias NervesHub.{Deployments, Devices, Products, ...}` is a code
smell (@jjcarstens can swear to this)

We have only `readability` and `software design suggestions` issues to
address, or ignore. These primarily consist of:
- `Modules should have a @moduledoc tag`
- `Nested modules could be aliased at the top of the invoking module.`
- and `Found a TODO tag in a comment`

I think we should address all of these, but we can instead leave these
for another PR.

---------

Co-authored-by: Jon Carstens <jon.carstens@smartrent.com>