Move all audit logging to templates (#1709)
Moves all audit logging to the template module.

One exception is in the function `Devices.update_device_with_audit/4`
where the audit log description is passed via the updating action (like
`tag_device`for example). I don't know if it's worth it adding this to
templates, since the templates themselves (the description) are defined
outside the templates module. Any input on this is welcomed!