Variant of the progress bar tweaks (#1816)
If you can't tell I really like the thin progress bar :D 


![Screenshot_20250123_095813](https://github.com/user-attachments/assets/6b6065f4-40d7-4825-9dbe-a0f2a055ec3f)

This uses the fix of the background from:
https://github.com/nerves-hub/nerves_hub_web/pull/1814

But it also uses the thin stripe.

For simplicity I cut the % display in the label. I think the label is
helpful but it was making the whole table jiggle and it would be a bunch
of faffing about to make it behave nice.