Remove audit and logger statement from extensions request check
Older versions of NervesHubLink would crash if an unknown messages was
pushed down over the socket. Since extensions required a new message,
this could fail badly on older devices and so a check was put in
place to verify the Device API version NHL was reporting to know
if it is safe to request extension info or not.

This removes the audit log _and_ logger statement that was happening
when this safe check would not pass meaning the device version was
too old to request. However, this happens with _every_ socket connection
and can produce incredible amount of noise.

If that level of info is needed, I suggest a future change which shows
the device version reported and what that means for support on the device
(i.e. in the Device settings page perhaps?)