Fix an extensions bug related to channels not sharing assigns
We don't know of a devices firmware metadata until it connects for the first time. When it does connect, the `DeviceSocket` will fetch the device and the `firmware_metadata` field will be nil. This device is the one shared with other channels.

In `Extensions.Health` we are using `socket.assigns.device.firmware_metadata` which is going to be nil during its first connection to NervesHub, and will continue to exit sending reports and creating Sentry errors.

This test replicates the scenario of a device connecting for the first time,  then its firmware metadata being updated, and then connecting to the health extension, exposing the error.

Sadly, in tests, channels share assigns, so this was much harder to replicate.