Fix an issue where `on_disconnect` was being called twice
This has also been seen in production. I believe it's related to channels being told to disconnect, and if a socket has two channels, we see two disconnection messages. Or it could be Phoenix/Bandit sending it twice for some reason?

Either way, this tracks when it was first called and storing it in the assigns.