Run one deployment orchestrator per deployment per cluster (#1865)
Currently we run one `Deployments.Orchestrator` per deployment per node,
it doesn't matter if the deployment is active or not, it will still be
run on each node.

If we have 100 deployments, and 10 nodes, then we have 100 orchestrators
per node, and 1000 overalls.

This architecture, specifically having an orchestrator for a deployment
run on every node, makes it very hard to ensure only a certain number of
concurrent updates are running at the same time. And it also makes it
hard to build out deployment workflows.

This PR uses `ProcessHub` to ensure one orchestrator is running per
deployment across the cluster. Additionally, it only runs orchestrators
that are set as active.

This ensures centralised management of a deployment, while using
`ProcessHub` to redistribute orchestrators to other nodes if a node is
to go offline.

The current implementation allows for us to switch between strategies.

If we commit to this path (after testing in the field), we can remove
the device registry, simplifying device connections. Plus, it will also
enable us to add support for other device transports and transport
setups, like a socket proxy or mqtt.

To run this locally, open up two terminals and run:
```elixir
make iex-server-clustered
```
and
```elixir
make iex-server-clustered num=2
```

The new distributed orchestrators are set to only run on **web** nodes,
which also builds up to us supporting a 'device proxy' setup.

---------

Co-authored-by: Nate Shoemaker <nshoes@users.noreply.github.com>
Co-authored-by: Nate Shoemaker <transition32@gmail.com>