Ensure available deployments match selected devices architecture and platform when batch-updating (#1912)
Previously, we displayed all deployments as options when batch-updating
devices. This is confusing because you could see deployments that
shouldn't be options. This PR now requires you to filter by platform
first before having the ability to bulk move devices into a deployment.

This change does create an extra step when moving devices to a
deployment but it's important to note this is an interim feature before
#1653 is finished.